Use block storage path fixture in storage test
